+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add the overview view to `bespoke` template ([#464](https://github.com/marp-team/marp-cli/issues/464), [#705](https://github.com/marp-team/marp-cli/pull/705) by [@mil-ad](https://github.com/mil-ad), [#710](https://github.com/marp-team/marp-cli/pull/710))
 - Set Liquid Glass icon while preview mode on macOS 26 Tahoe ([#688](https://github.com/marp-team/marp-cli/pull/688))
 
 ### Changed